+++ CHANGES.md
@@ -24,6 +24,7 @@
 * Add special case of training heuristic with random policy when using pathfinding algorithm that uses a policy and not training a policy
 * Add timing of getting supervised data to supervised pathfinding
 * Add abstract method for loss info for training policy
+* Add layer norm to resnet_fc
 
 ## 0.2.1
 * Consolidate search: Beam search -> special cases: greedy_policy, graph search -> special cases: batch weighted A* search, batch weighted Q* search